requirements.txt
numpy
pandas
scikit-learn
matplotlib
umap-learn
